Add files via upload
Having decided to immerse myself in the world of mobile applications, I accepted the challenge and created my first project, Pretec. I have poured my heart and soul into this application, striving to adhere to all the principles of UX/UI design and perfecting it every day.
Pretec represents not only the result of my work, but also the journey of my development as a mobile designer. I paid careful attention to every detail, animation, and UX/UI concept to create a user-friendly and engaging experience.
This project has been a test not only of my skills but also an opportunity for continuous improvement. I am proud of each step I have taken in creating Pretec, and I believe it can serve as an inspiration to others and a demonstration of my dedication to high-quality mobile app design.